moved scripts to src/
